mypyの型チェック対象をパッケージとして実行する
