+++ mypy.ini
@@ -1,0 +1,4 @@
+[mypy]
+
+[mypy-python_api_template.*]
+ignore_missing_imports = True
